requirements.txt
Flask==0.8
gntp==0.5.1
pyyaml==3.10
